x64版でXIの書き出しが上手く行ってなかったのを修正
git-svn-id: http://picopicose.com/svn/C700@121 38616c85-8a89-495d-9829-d15a903298fa

+++ XIFile.cpp
@@ -167,7 +167,7 @@ bool XIFile::SetDataFromChip( const Chip700Generator *chip, int targetProgram, d
 			bool	existSrcFile = false;
 			if ( chip->getVP(ismp)->sourceFile[0] ) {
 				AudioFile::InstData	inst;
-				long		numSamples;
+				int		numSamples;
 				
 				AudioFile	origFile(chip->getVP(ismp)->sourceFile,false);
 				origFile.Load();
@@ -217,7 +217,7 @@ bool XIFile::SetDataFromChip( const Chip700Generator *chip, int targetProgram, d
 	for (int ismp=start_prg; ismp<=end_prg; ismp++) {
 		if ( chip->getVP(ismp)->brr.data != NULL && chip->getVP(ismp)->bank == selectBank ) {
 			short	*wavedata;
-			long	numSamples;
+			int	numSamples;
 			bool	existSrcFile = false;	//元ファイルが存在するか？
 			
 			if ( chip->getVP(ismp)->sourceFile[0] ) {
